Refactor UserDetailScreen to use ViewModel and remove dummy data; add UserDetailViewModel for state management
